requirements.txt
# バックエンドフレームワークとコア機能
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# データベース
mysqlclient==2.2.0

# 認証・セキュリティ
PyJWT==2.8.0
python-dotenv==1.0.0

# シリアライズ・バリデーション
marshmallow==3.20.1

# 開発・テスト用
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# 追加されたWerkzeugバージョン
Werkzeug==2.3.7 

# notion_sync.py用ライブラリ
requests
pandas